backend/requirements.txt
flask
requests
flask-cors
google.generativeai
python-dotenv   
